requirements.txt
streamlit>=1.32.0
tensorflow>=2.13.0,<2.14
pillow>=9.0.0
scikit-learn>=1.0.2
numpy>=1.21.0
pandas>=1.3.0
openpyxl>=3.0.10
xlsxwriter>=3.0.9
